# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];


export const ProjectStatus = {
  active: 'active',
  archived: 'archived',
} as const;

export type ProjectVisibility = typeof ProjectVisibility[keyof typeof ProjectVisibility];


export const ProjectVisibility = {
  private: 'private',
  public: 'public',
} as const;

export interface Project {
  id: string;
  name: string;
  slug: string;
  description: string;
  language: string;
  status: ProjectStatus;
  visibility: ProjectVisibility;
  updatedAt: string;
  createdAt: string;
  accent: string;
}

export type ProjectInputVisibility = typeof ProjectInputVisibility[keyof typeof ProjectInputVisibility];


export const ProjectInputVisibility = {
  private: 'private',
  public: 'public',
} as const;

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  language: string;
  visibility?: ProjectInputVisibility;
}

export type ProjectUpdateStatus = typeof ProjectUpdateStatus[keyof typeof ProjectUpdateStatus];


export const ProjectUpdateStatus = {
  active: 'active',
  archived: 'archived',
} as const;

export type ProjectUpdateVisibility = typeof ProjectUpdateVisibility[keyof typeof ProjectUpdateVisibility];


export const ProjectUpdateVisibility = {
  private: 'private',
  public: 'public',
} as const;

export interface ProjectUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  language?: string;
  status?: ProjectUpdateStatus;
  visibility?: ProjectUpdateVisibility;
}

export type ActivityType = typeof ActivityType[keyof typeof ActivityType];


export const ActivityType = {
  project_created: 'project_created',
  project_updated: 'project_updated',
  project_archived: 'project_archived',
  deployment: 'deployment',
  collaborator: 'collaborator',
} as const;

export interface Activity {
  id: string;
  type: ActivityType;
  title: string;
  detail: string;
  /** @nullable */
  projectName: string | null;
  createdAt: string;
}

export type DashboardStats = {
  totalProjects: number;
  activeProjects: number;
  deployments: number;
  storageUsed: string;
};

export interface Dashboard {
  greeting: string;
  stats: DashboardStats;
  featuredProject: Project | null;
  recentActivity: Activity[];
}

export type CapabilityStatus = typeof CapabilityStatus[keyof typeof CapabilityStatus];


export const CapabilityStatus = {
  available: 'available',
  coming_soon: 'coming_soon',
} as const;

export interface Capability {
  id: string;
  label: string;
  description: string;
  status: CapabilityStatus;
  category: string;
}

export type ListProjectsParams = {
q?: string;
status?: ListProjectsStatus;
};

export type ListProjectsStatus = typeof ListProjectsStatus[keyof typeof ListProjectsStatus];


export const ListProjectsStatus = {
  active: 'active',
  archived: 'archived',
  all: 'all',
} as const;

export type ListActivityParams = {
/**
 * @minimum 1
 * @maximum 50
 */
limit?: number;
};