# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: workspace
    description: Workspace overview and capability discovery
  - name: projects
    description: Project lifecycle operations
  - name: activity
    description: Workspace activity feed
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /dashboard:
    get:
      operationId: getDashboard
      tags: [workspace]
      summary: Get the workspace dashboard
      description: Returns the current workspace summary, project metrics, and suggested next actions.
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
  /capabilities:
    get:
      operationId: getCapabilities
      tags: [workspace]
      summary: List nabeen capabilities
      description: Returns the product capabilities currently available or planned in the workspace.
      responses:
        "200":
          description: Capability list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Capability"
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, archived, all]
            default: all
      responses:
        "200":
          description: Project list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Project not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Project not found
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "204":
          description: Project deleted
        "404":
          description: Project not found
  /activity:
    get:
      operationId: listActivity
      tags: [activity]
      summary: List recent activity
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        "200":
          description: Activity feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
components:
  parameters:
    ProjectId:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        language:
          type: string
        status:
          type: string
          enum: [active, archived]
        visibility:
          type: string
          enum: [private, public]
        updatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        accent:
          type: string
      required:
        - id
        - name
        - slug
        - description
        - language
        - status
        - visibility
        - updatedAt
        - createdAt
        - accent
    ProjectInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        language:
          type: string
        visibility:
          type: string
          enum: [private, public]
      required:
        - name
        - language
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        language:
          type: string
        status:
          type: string
          enum: [active, archived]
        visibility:
          type: string
          enum: [private, public]
    Activity:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [project_created, project_updated, project_archived, deployment, collaborator]
        title:
          type: string
        detail:
          type: string
        projectName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - title
        - detail
        - projectName
        - createdAt
    Dashboard:
      type: object
      properties:
        greeting:
          type: string
        stats:
          type: object
          properties:
            totalProjects:
              type: integer
            activeProjects:
              type: integer
            deployments:
              type: integer
            storageUsed:
              type: string
          required: [totalProjects, activeProjects, deployments, storageUsed]
        featuredProject:
          anyOf:
            - $ref: "#/components/schemas/Project"
            - type: "null"
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/Activity"
      required:
        - greeting
        - stats
        - featuredProject
        - recentActivity
    Capability:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [available, coming_soon]
        category:
          type: string
      required: [id, label, description, status, category]