# lib/api-zod/src/generated/types/capability.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CapabilityStatus } from './capabilityStatus';

export interface Capability {
  id: string;
  label: string;
  description: string;
  status: CapabilityStatus;
  category: string;
}