# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Activity } from './activity';
import type { DashboardStats } from './dashboardStats';
import type { Project } from './project';

export interface Dashboard {
  greeting: string;
  stats: DashboardStats;
  featuredProject: Project | null;
  recentActivity: Activity[];
}