# lib/api-zod/src/generated/types/projectUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectUpdateStatus } from './projectUpdateStatus';
import type { ProjectUpdateVisibility } from './projectUpdateVisibility';

export interface ProjectUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  language?: string;
  status?: ProjectUpdateStatus;
  visibility?: ProjectUpdateVisibility;
}